requirements.txt
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
jmespath==1.0.1
numpy==2.2.6
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
s3transfer==0.13.0
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
